Split multi-type files into one type per file
Moves the four endpoint filters out of SmidgeEndpointFilters.cs and the
integration test fixtures out of SmidgeCacheEndpointTests.cs into
individual files, and extracts the nested test helper types
(TempFolder, pre-processor stubs) into their own files. No behaviour
changes; purely a file layout refactor.
