+++ test/Smidge.Integration.Tests/InMemoryCacheAppFixture.cs
@@ -1,0 +1,7 @@
+namespace Smidge.Integration.Tests
+{
+    public sealed class InMemoryCacheAppFixture : SmidgeAppFixture
+    {
+        public InMemoryCacheAppFixture() : base(inMemory: true) { }
+    }
+}
