+++ test/Smidge.Integration.Tests/InMemoryCacheEndpointTests.cs
@@ -1,0 +1,9 @@
+using Xunit;
+
+namespace Smidge.Integration.Tests
+{
+    public sealed class InMemoryCacheEndpointTests : SmidgeEndpointTestsBase, IClassFixture<InMemoryCacheAppFixture>
+    {
+        public InMemoryCacheEndpointTests(InMemoryCacheAppFixture fixture) : base(fixture) { }
+    }
+}
