+++ test/Smidge.Integration.Tests/PhysicalCacheAppFixture.cs
@@ -1,0 +1,7 @@
+namespace Smidge.Integration.Tests
+{
+    public sealed class PhysicalCacheAppFixture : SmidgeAppFixture
+    {
+        public PhysicalCacheAppFixture() : base(inMemory: false) { }
+    }
+}
