+++ test/Smidge.Integration.Tests/PhysicalCacheEndpointTests.cs
@@ -1,0 +1,9 @@
+using Xunit;
+
+namespace Smidge.Integration.Tests
+{
+    public sealed class PhysicalCacheEndpointTests : SmidgeEndpointTestsBase, IClassFixture<PhysicalCacheAppFixture>
+    {
+        public PhysicalCacheEndpointTests(PhysicalCacheAppFixture fixture) : base(fixture) { }
+    }
+}
