+++ test/Smidge.Tests/CacheFileSystemTests.cs
@@ -54,28 +54,5 @@ namespace Smidge.Tests
 
             Assert.Throws<FileNotFoundException>(() => fs.GetRequiredFileInfo("does-not-exist.css"));
         }
-
-        private sealed class TempFolder : IDisposable
-        {
-            public TempFolder()
-            {
-                Path = System.IO.Path.Combine(System.IO.Path.GetTempPath(), "smidge-tests-" + Guid.NewGuid().ToString("N"));
-                Directory.CreateDirectory(Path);
-            }
-
-            public string Path { get; }
-
-            public void Dispose()
-            {
-                try
-                {
-                    Directory.Delete(Path, true);
-                }
-                catch
-                {
-                    // best effort cleanup
-                }
-            }
-        }
     }
 }
